Updated fonts throughout the page to use 'Poppins' loaded via Google CDN.
In order to improve the appearance of the page, Poppins is loaded with a serif fallback font. This currently causes the dread "FOUT" effect, aka the Flash Of Unstyled Text. It is incredibly brief but it may be worth revisiting the issue later to avoid loading any text until the font has been loaded from Google's CDN.