Project cards added
This commit adds "project cards" to the projects section of the home page.
These cards reveal details when hovered (on tablet screens and higher)
that link to GitHub while also providing details about the project.

The card component itself is nearly done outside of not being responsive.
Some information inside of ProjectsContainer needs to be updated still,
but that will have to wait until this portfolio is done (in vanilla React).

This commit resolves the "about me" issue.