Refactor: refactored application to use TypeScript
For the sake of rapid development while learning Next.js the app
was originally written in JavaScript. Now that most of the app has
been constructed, it makes sense to migrate to TypeScript before
continuing to add functionality such as GraphQL.

Note, in some instances there are interfaces and types relying on
"any" types which may want to be refactored further later. Effort
has been made to ensure that this does not occur in components that
rely on client side data, but this may need to be reviewed when
GraphQL and data-fetching is added later.

This resolves the "typescript" issue from the backlog.