Added basic testing setup
This commit contains some of the introductory configuration for jest,
husky, and lint-staged. The goal is to make it easier to automate
all future testing with this setup, so that all committed code is
linted and tested prior to a commit going through.

Note that the tests themselves are not fully completed. Some basic unit
tests have been added but there are no integration tests or e2e tests
yet.

This resolves the "testing" issue from the backlog.