AWS Deployment and more
This commit deploys the application to AWS using serverless-next.js
which hosts static content in an S3 bucket while serving SSR content
from an AWS Lambda. In this case, all content is SSG so only an S3
bucket is used. From there, content id delivered via CloudFront.

Other changes include some SEO tweaks to the head tag, an update to
the About Me section, and some minor corrections to the copy for
the ProjectsContainer.